+++ CoreLayer/Entities/Product.cs
@@ -8,14 +8,14 @@ namespace Domain.Entities
 {
     public class Product
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Brand { get; set; }
         public string Name { get; set; }
         public string Category { get; set; }
         public string AffiliateURL { get; set; }
         public string ImageURL { get; set; }
         public decimal? Price { get; set; }
-        public int? PartnerID { get; set; }
+        public long? PartnerID { get; set; }
         public bool IsActive { get; set; }
 
         public Partner? Partner { get; set; }
